Corregida librería HAL en el archivo stm32f1xx_hal_rcc_ex.c iniciar RTC
Agregado casting (uint32_t) en línea 584 de stm32f1xx_hal_rcc_ex.c para
que al retornar se retorne el valor correcto (antes retornaba cero)
pueda iniciar el RTC.

Antes: return (frequency)
Después: return (uint32_t) (frequency)

+++ Esclava/Drivers/STM32F1xx_HAL_Driver/Src/stm32f1xx_hal_rcc_ex.c
@@ -581,7 +581,7 @@ uint32_t HAL_RCCEx_GetPeriphCLKFreq(uint32_t PeriphClk)
       break;
     }
   }
-  return (frequency);
+  return (uint32_t) frequency;
 }
 
 /**
